backend/requirements.txt
cloudinary
python-dotenv
pymongo[srv]
bcrypt
fastapi
uvicorn[standard]
python-multipart
Pillow
requests
openai